continue lib api cleanup

+++ libfreeipmi/src/api/ipmi-ssif-driver-api.c
@@ -81,7 +81,7 @@ ipmi_ssif_cmd_api (ipmi_ctx_t ctx,
 
   {
     uint8_t *pkt;
-    uint32_t pkt_len;
+    unsigned int pkt_len;
     int32_t hdr_len, cmd_len, send_len;
 
     if ((hdr_len = fiid_template_len_bytes (tmpl_hdr_kcs)) < 0)
@@ -132,8 +132,8 @@ ipmi_ssif_cmd_api (ipmi_ctx_t ctx,
 
   {
     uint8_t *pkt;
-    uint32_t pkt_len;
-    ssize_t read_len;
+    unsigned int pkt_len;
+    int read_len;
     int32_t hdr_len, cmd_len;
     fiid_field_t *tmpl = NULL;
     int rv = -1;
@@ -204,9 +204,9 @@ ipmi_ssif_cmd_raw_api (ipmi_ctx_t ctx,
                        size_t buf_rs_len)
 {
   uint8_t *pkt = NULL;
-  uint32_t pkt_len;
+  unsigned int pkt_len;
   uint8_t *readbuf = NULL;
-  int32_t bytes_read = 0;
+  int bytes_read = 0;
   int32_t hdr_len;
   int32_t rv = -1;
